+++ src/main/java/com/cleanroommc/modularui/factory/ClientGUI.java
@@ -1,5 +1,6 @@
 package com.cleanroommc.modularui.factory;
 
+import com.cleanroommc.modularui.screen.ContainerCustomizer;
 import com.cleanroommc.modularui.screen.JeiSettingsImpl;
 import com.cleanroommc.modularui.screen.ModularScreen;
 
@@ -9,6 +10,7 @@ import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 /**
  * Helper class to open client only GUIs. This class is safe to use inside a Modular GUI.
@@ -27,7 +29,7 @@ public class ClientGUI {
      * @param screen new modular screen
      */
     public static void open(@NotNull ModularScreen screen) {
-        open(screen, new JeiSettingsImpl());
+        open(screen, new JeiSettingsImpl(), null);
     }
 
     /**
@@ -38,7 +40,30 @@ public class ClientGUI {
      * @param jeiSettings custom jei settings
      */
     public static void open(@NotNull ModularScreen screen, @NotNull JeiSettingsImpl jeiSettings) {
-        GuiManager.openScreen(screen, jeiSettings);
+        GuiManager.openScreen(screen, jeiSettings, null);
+    }
+
+    /**
+     * Opens a modular screen on the next client tick with custom jei settings.
+     * It needs to be opened in next tick, because we might break the current GUI if we open it now.
+     *
+     * @param screen              new modular screen
+     * @param containerCustomizer container customizer
+     */
+    public static void open(@NotNull ModularScreen screen, @NotNull ContainerCustomizer containerCustomizer) {
+        GuiManager.openScreen(screen, new JeiSettingsImpl(), containerCustomizer);
+    }
+
+    /**
+     * Opens a modular screen on the next client tick with custom jei settings.
+     * It needs to be opened in next tick, because we might break the current GUI if we open it now.
+     *
+     * @param screen              new modular screen
+     * @param jeiSettings         custom jei settings
+     * @param containerCustomizer container customizer
+     */
+    public static void open(@NotNull ModularScreen screen, @NotNull JeiSettingsImpl jeiSettings, @Nullable ContainerCustomizer containerCustomizer) {
+        GuiManager.openScreen(screen, jeiSettings, containerCustomizer);
     }
 
     /**
