+++ src/main/java/com/cleanroommc/modularui/screen/ModularContainer.java
@@ -13,25 +13,22 @@ import com.cleanroommc.bogosorter.api.ISortingContextBuilder;
 
 import net.minecraft.client.Minecraft;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.entity.player.InventoryPlayer;
 import net.minecraft.inventory.ClickType;
 import net.minecraft.inventory.Container;
+import net.minecraft.inventory.IInventory;
 import net.minecraft.inventory.Slot;
 import net.minecraft.item.ItemStack;
+import net.minecraft.world.World;
 import net.minecraftforge.fml.common.Optional;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
-import net.minecraftforge.items.ItemHandlerHelper;
 
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Objects;
+import java.util.*;
 
 @Optional.Interface(modid = ModularUI.BOGO_SORT, iface = "com.cleanroommc.bogosorter.api.ISortableContainer")
 public class ModularContainer extends Container implements ISortableContainer {
@@ -48,22 +45,26 @@ public class ModularContainer extends Container implements ISortableContainer {
     private boolean init = true;
     private final List<ModularSlot> slots = new ArrayList<>();
     private final List<ModularSlot> shiftClickSlots = new ArrayList<>();
-
-    private static final int DROP_TO_WORLD = -999;
-    private static final int LEFT_MOUSE = 0;
-    private static final int RIGHT_MOUSE = 1;
+    private ContainerCustomizer containerCustomizer;
 
     public ModularContainer(EntityPlayer player, PanelSyncManager panelSyncManager, String mainPanelName) {
         this.player = player;
         this.syncManager = new ModularSyncManager(this);
         this.syncManager.construct(mainPanelName, panelSyncManager);
+        this.containerCustomizer = panelSyncManager.getContainerCustomizer();
+        if (this.containerCustomizer == null) {
+            this.containerCustomizer = new ContainerCustomizer();
+        }
+        this.containerCustomizer.initialize(this);
         sortShiftClickSlots();
     }
 
     @SideOnly(Side.CLIENT)
-    public ModularContainer() {
+    public ModularContainer(ContainerCustomizer containerCustomizer) {
         this.player = Minecraft.getMinecraft().player;
         this.syncManager = null;
+        this.containerCustomizer = containerCustomizer != null ? containerCustomizer : new ContainerCustomizer();
+        this.containerCustomizer.initialize(this);
     }
 
     public ContainerAccessor acc() {
@@ -76,6 +77,7 @@ public class ModularContainer extends Container implements ISortableContainer {
         if (this.syncManager != null) {
             this.syncManager.onClose();
         }
+        this.containerCustomizer.onContainerClosed();
     }
 
     @Override
@@ -166,181 +168,42 @@ public class ModularContainer extends Container implements ISortableContainer {
         return player;
     }
 
+    public ModularSlot getModularSlot(int index) {
+        return this.slots.get(index);
+    }
+
+    public List<ModularSlot> getShiftClickSlots() {
+        return Collections.unmodifiableList(this.shiftClickSlots);
+    }
+
     @Override
     public boolean canInteractWith(@NotNull EntityPlayer playerIn) {
         return true;
     }
 
     @Override
-    public @NotNull ItemStack slotClick(int slotId, int mouseButton, @NotNull ClickType clickTypeIn, EntityPlayer player) {
-        ItemStack returnable = ItemStack.EMPTY;
-        InventoryPlayer inventoryplayer = player.inventory;
-
-        if (clickTypeIn == ClickType.QUICK_CRAFT || acc().getDragEvent() != 0) {
-            return super.slotClick(slotId, mouseButton, clickTypeIn, player);
-        }
-
-        if ((clickTypeIn == ClickType.PICKUP || clickTypeIn == ClickType.QUICK_MOVE) &&
-                (mouseButton == LEFT_MOUSE || mouseButton == RIGHT_MOUSE)) {
-            if (slotId == DROP_TO_WORLD) {
-                if (!inventoryplayer.getItemStack().isEmpty()) {
-                    if (mouseButton == LEFT_MOUSE) {
-                        player.dropItem(inventoryplayer.getItemStack(), true);
-                        inventoryplayer.setItemStack(ItemStack.EMPTY);
-                    }
-
-                    if (mouseButton == RIGHT_MOUSE) {
-                        player.dropItem(inventoryplayer.getItemStack().splitStack(1), true);
-                    }
-                }
-                return inventoryplayer.getItemStack();
-            }
-
-            if (slotId < 0) return ItemStack.EMPTY;
-
-            if (clickTypeIn == ClickType.QUICK_MOVE) {
-                Slot fromSlot = getSlot(slotId);
-
-                if (!fromSlot.canTakeStack(player)) {
-                    return ItemStack.EMPTY;
-                }
-
-                returnable = this.transferStackInSlot(player, slotId);
-            } else {
-                Slot clickedSlot = getSlot(slotId);
-
-                ItemStack slotStack = clickedSlot.getStack();
-                ItemStack heldStack = inventoryplayer.getItemStack();
-
-                if (slotStack.isEmpty()) {
-                    if (!heldStack.isEmpty() && clickedSlot.isItemValid(heldStack)) {
-                        int stackCount = mouseButton == LEFT_MOUSE ? heldStack.getCount() : 1;
-
-                        if (stackCount > clickedSlot.getItemStackLimit(heldStack)) {
-                            stackCount = clickedSlot.getItemStackLimit(heldStack);
-                        }
-
-                        clickedSlot.putStack(heldStack.splitStack(stackCount));
-                    }
-                } else if (clickedSlot.canTakeStack(player)) {
-                    if (heldStack.isEmpty() && !slotStack.isEmpty()) {
-                        int toRemove = mouseButton == LEFT_MOUSE ? slotStack.getCount() : (slotStack.getCount() + 1) / 2;
-                        inventoryplayer.setItemStack(slotStack.splitStack(toRemove));
-                        clickedSlot.putStack(slotStack);
-
-                        clickedSlot.onTake(player, inventoryplayer.getItemStack());
-                    } else if (clickedSlot.isItemValid(heldStack)) {
-                        if (slotStack.getItem() == heldStack.getItem() &&
-                                slotStack.getMetadata() == heldStack.getMetadata() &&
-                                ItemStack.areItemStackTagsEqual(slotStack, heldStack)) {
-                            int stackCount = mouseButton == LEFT_MOUSE ? heldStack.getCount() : 1;
-
-                            if (stackCount > clickedSlot.getItemStackLimit(heldStack) - slotStack.getCount()) {
-                                stackCount = clickedSlot.getItemStackLimit(heldStack) - slotStack.getCount();
-                            }
-
-                            if (stackCount > heldStack.getMaxStackSize() - slotStack.getCount()) {
-                                stackCount = heldStack.getMaxStackSize() - slotStack.getCount();
-                            }
-
-                            heldStack.shrink(stackCount);
-                            slotStack.grow(stackCount);
-                            clickedSlot.putStack(slotStack);
-
-                        } else if (heldStack.getCount() <= clickedSlot.getItemStackLimit(heldStack)) {
-                            clickedSlot.putStack(heldStack);
-                            inventoryplayer.setItemStack(slotStack);
-                        }
-                    } else if (slotStack.getItem() == heldStack.getItem() &&
-                            heldStack.getMaxStackSize() > 1 &&
-                            (!slotStack.getHasSubtypes() || slotStack.getMetadata() == heldStack.getMetadata()) &&
-                            ItemStack.areItemStackTagsEqual(slotStack, heldStack) && !slotStack.isEmpty()) {
-                        int stackCount = slotStack.getCount();
-
-                        if (stackCount + heldStack.getCount() <= heldStack.getMaxStackSize()) {
-                            heldStack.grow(stackCount);
-                            slotStack = clickedSlot.decrStackSize(stackCount);
-
-                            if (slotStack.isEmpty()) {
-                                clickedSlot.putStack(ItemStack.EMPTY);
-                            }
-
-                            clickedSlot.onTake(player, inventoryplayer.getItemStack());
-                        }
-                    }
-                }
-                clickedSlot.onSlotChanged();
-            }
-            this.detectAndSendChanges();
-            return returnable;
-        }
-
-        return super.slotClick(slotId, mouseButton, clickTypeIn, player);
+    public @NotNull ItemStack slotClick(int slotId, int mouseButton, @NotNull ClickType clickTypeIn, @NotNull EntityPlayer player) {
+        return this.containerCustomizer.slotClick(slotId, mouseButton, clickTypeIn, player);
     }
-
 
     @Override
     public @NotNull ItemStack transferStackInSlot(@NotNull EntityPlayer playerIn, int index) {
-        ModularSlot slot = this.slots.get(index);
-        if (!slot.isPhantom()) {
-            ItemStack stack = slot.getStack();
-            if (!stack.isEmpty()) {
-                ItemStack remainder = transferItem(slot, stack.copy());
-                if (remainder.isEmpty()) stack = ItemStack.EMPTY;
-                else stack.setCount(remainder.getCount());
-                slot.putStack(stack);
-                return ItemStack.EMPTY;
-            }
-        }
-        return ItemStack.EMPTY;
+        return this.containerCustomizer.transferStackInSlot(playerIn, index);
     }
 
-    protected ItemStack transferItem(ModularSlot fromSlot, ItemStack fromStack) {
-        @Nullable SlotGroup fromSlotGroup = fromSlot.getSlotGroup();
-        for (ModularSlot toSlot : this.shiftClickSlots) {
-            SlotGroup slotGroup = Objects.requireNonNull(toSlot.getSlotGroup());
-            if (slotGroup != fromSlotGroup && toSlot.isEnabled() && toSlot.isItemValid(fromStack)) {
-                ItemStack toStack = toSlot.getStack().copy();
-                if (toSlot.isPhantom()) {
-                    if (toStack.isEmpty() || (ItemHandlerHelper.canItemStacksStack(fromStack, toStack) && toStack.getCount() < toSlot.getItemStackLimit(toStack))) {
-                        toSlot.putStack(fromStack.copy());
-                        return fromStack;
-                    }
-                } else if (ItemHandlerHelper.canItemStacksStack(fromStack, toStack)) {
-                    int j = toStack.getCount() + fromStack.getCount();
-                    int maxSize = Math.min(toSlot.getSlotStackLimit(), fromStack.getMaxStackSize());
+    @Override
+    public boolean canMergeSlot(@NotNull ItemStack stack, @NotNull Slot slotIn) {
+        return this.containerCustomizer.canMergeSlot(stack, slotIn);
+    }
 
-                    if (j <= maxSize) {
-                        fromStack.setCount(0);
-                        toStack.setCount(j);
-                        toSlot.putStack(toStack);
-                    } else if (toStack.getCount() < maxSize) {
-                        fromStack.shrink(maxSize - toStack.getCount());
-                        toStack.setCount(maxSize);
-                        toSlot.putStack(toStack);
-                    }
+    @Override
+    protected boolean mergeItemStack(@NotNull ItemStack stack, int startIndex, int endIndex, boolean reverseDirection) {
+        return this.containerCustomizer.mergeItemStack(stack, startIndex, endIndex, reverseDirection);
+    }
 
-                    if (fromStack.isEmpty()) {
-                        return fromStack;
-                    }
-                }
-            }
-        }
-        for (ModularSlot emptySlot : this.shiftClickSlots) {
-            ItemStack itemstack = emptySlot.getStack();
-            SlotGroup slotGroup = Objects.requireNonNull(emptySlot.getSlotGroup());
-            if (slotGroup != fromSlotGroup && emptySlot.isEnabled() && itemstack.isEmpty() && emptySlot.isItemValid(fromStack)) {
-                if (fromStack.getCount() > emptySlot.getSlotStackLimit()) {
-                    emptySlot.putStack(fromStack.splitStack(emptySlot.getSlotStackLimit()));
-                } else {
-                    emptySlot.putStack(fromStack.splitStack(fromStack.getCount()));
-                }
-                if (fromStack.getCount() < 1) {
-                    break;
-                }
-            }
-        }
-        return fromStack;
+    @Override
+    protected void clearContainer(@NotNull EntityPlayer playerIn, @NotNull World worldIn, @NotNull IInventory inventoryIn) {
+        this.containerCustomizer.clearContainer(playerIn, worldIn, inventoryIn);
     }
 
     @Override
